db: Change the type of the some blockNo columns
Some table colums that help the `blockNo` field had a `Word64` type
when they should all be the same type (Int64 in Haskell and bigint
in SQL).

+++ cardano-db/test/Test/IO/Cardano/Db/Insert.hs
@@ -116,7 +116,7 @@ adaPotsZero (BlockNo blkNo) =
     , adaPotsUtxo = DbLovelace 0
     , adaPotsDeposits = DbLovelace 0
     , adaPotsFees = DbLovelace 0
-    , adaPotsBlockNo = blkNo
+    , adaPotsBlockNo = fromIntegral blkNo
     }
 
 mkHash :: Int -> Char -> ByteString
